requirements.txt
numpy==1.26.4
opencv-python
Pillow
tqdm
matplotlib
scikit-learn
segmentation-models-pytorch
albumentations
pyyaml
tensorboard
